backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
Pillow==9.0.0
djoser==2.2.3

djangorestframework-simplejwt
django-filter~=2.4.0
drf-yasg
drf-extra-fields
shortuuid